+++ tests/test_sessions.py
@@ -1,5 +1,4 @@
-import random
-
+import pytest
 import sqlalchemy as sa
 from sqlalchemy.orm import sessionmaker
 
@@ -10,7 +9,8 @@ def test_default_session_scoping(app, db):
     class FOOBar(db.Model):
         id = db.Column(db.Integer, primary_key=True)
 
-    db.create_all()
+    with app.app_context():
+        db.create_all()
 
     with app.test_request_context():
         fb = FOOBar()
@@ -19,17 +19,32 @@ def test_default_session_scoping(app, db):
 
 
 def test_session_scoping_changing(app):
-    db = SQLAlchemy(app, session_options={"scopefunc": random.random})
+    count = 0
+
+    def scope():
+        nonlocal count
+        count += 1
+        return count
+
+    db = SQLAlchemy(app, session_options={"scopefunc": scope})
 
     class Example(db.Model):
         id = db.Column(db.Integer, primary_key=True)
 
-    db.create_all()
-    fb = Example()
-    db.session.add(fb)
-    assert fb not in db.session  # because a new scope is generated on each call
+    with app.app_context():
+        db.create_all()
+        fb = Example()
+        db.session.add(fb)
+        assert fb not in db.session  # because a new scope is generated on each call
+        assert count == 2
+
+        for session in db.session.registry.registry.values():
+            session.close()
+
+        db.session.registry.registry.clear()
 
 
+@pytest.mark.usefixtures("app_ctx")
 def test_insert_update_delete(db):
     # Ensure _SignalTrackingMapperExtension doesn't croak when
     # faced with a vanilla SQLAlchemy session. Verify that